Team Image Animations, SEO Wording, Image Fix
Overlay Animation added to each of the team images, as well as centred name text because it looks better.

Added Casing Design Software to a couple of areas as we discussed.

Changes images from /images/ to \images (as you did for OCTG images for me). Now, ALL images work when I fetch from Github and hopefully should work when submitted.