Full page now designed
Future descriptions to have more information.

We have the main header, a hero image with no navigation. No navigation is reccomended, as a few websites and marketers suggested it is distracting for a landing page. We need them to keep their eyes on the prize. Buttons very visible for conversion optimization.

Following that we have client logos, showing who uses the software. These need to be checked, either  with Paul or Natalie.

Then we have "why stringnosis", which is 4 quarters with text from the main website. It's good information about why to pick stringnosis.

We have the features sections, with modals popping up to the left. We need to fill in the descriptions, which I will need Pete help with.

We have a free trial/demo section. Demo section adds text to subject and body which will direct to support@stringnosis.com. Free trial targets _blank to open a new tab for creating an account.

In between sections we have some sales messaging.

Before the footer, we give people the option to visit the main website now that they have had an intro.

Then the footer, which is normal but I have fixed somethings to make it look nicer. Removed OCTG as discontinuing?