Made Edits after meeting with Pete
Made text on hero image bolder and added colour to the buttons.

Darkened the background for colours to stand out more.

Removed delivered by blade energy as it didnt fit anywhere.

Added link to stringnosis online mention in the "why" section, as well as a link to academy for where it says it's easy to learn.

Added descriptions to all of the modals from pre-existing information and flyers.

Added 1 hour to the demos description and 2 minutes for login time for the trial.

Fixed mobile layout issue with buttons on the free trial/demo sections.